Fixed hieracrchy of submitted files. See Piazza post @63
