+++ dominion/badTestDrawCard.c
@@ -4,8 +4,6 @@
 #include <stdio.h>
 #include <assert.h>
 #include "rngs.h"
-#include <math.h> //Added to silence warnings
-#include <stdlib.h> //Added to silence warnings
 
 #define DEBUG 0
 #define NOISY_TEST 1
@@ -14,23 +12,14 @@ int checkDrawCard(int p, struct gameState *post) {
   int r;
     
   r = drawCard (p, post);
-    
-    assert(r == 0); //Added to silence warnings
-    
-    return 0;
 }
 
 int main () {
-    
-    //Removed r, deskCount, discardCount, and handCount to silence warnings.
-  int i, n, p;
-    
 
-    //Removed to silene warnings
-    /*
+  int i, n, r, p, deckCount, discardCount, handCount;
+
   int k[10] = {adventurer, council_room, feast, gardens, mine,
 	       remodel, smithy, village, baron, great_hall};
-    */
 
   struct gameState G;
