+++ dominion/dominion.c
@@ -5,137 +5,6 @@
 #include <math.h>
 #include <stdlib.h>
 
-//REFACTOR Section
-//The following 5 functions are the refactored card statements orginally found in the cardEffect() switch statements.
-void adventurerRefactor(int drawntreasure, struct gameState *state, int currentPlayer, int temphand[], int z)
-{
-    int cardDrawn; //Assigned within refactored function
-    
-    while(drawntreasure<2){
-        if (state->deckCount[currentPlayer] <1){//if the deck is empty we need to shuffle discard and add to deck
-            shuffle(currentPlayer, state);
-        }
-        drawCard(currentPlayer, state);
-        cardDrawn = state->hand[currentPlayer][state->handCount[currentPlayer]-1];//top card of hand is most recently drawn card.
-        //*BUG INTRODUCED*
-        if (cardDrawn == copper || cardDrawn == silver || cardDrawn == gold)
-        {
-            if(cardDrawn == copper)
-            {
-                drawntreasure--;
-            }
-            else
-            {
-                drawntreasure++;
-            }
-            
-        }
-        
-        else{
-            temphand[z]=cardDrawn;
-            state->handCount[currentPlayer]--; //this should just remove the top card (the most recently drawn one).
-            z++;
-        }
-    }
-    while(z-1>=0){
-        state->discard[currentPlayer][state->discardCount[currentPlayer]++]=temphand[z-1]; // discard all cards in play that have been drawn
-        z=z-1;
-    }
-}
-
-void smithyRefactor(int currentPlayer, struct gameState *state, int i, int handPos)
-{
-    //+5 Cards
-    //*BUG INTRODUCED*
-    for (i = 0; i < 5; i++)
-    {
-        drawCard(currentPlayer, state);
-    }
-    
-    //discard card from hand
-    discardCard(handPos, currentPlayer, state, 0);
-}
-
-void feastRefactor(int currentPlayer, struct gameState *state, int i, int temphand[], int x, int choice1)
-{
-    //gain card with cost up to 5
-    //Backup hand
-    for (i = 0; i <= state->handCount[currentPlayer]; i++){
-        temphand[i] = state->hand[currentPlayer][i];//Backup card
-        state->hand[currentPlayer][i] = -1;//Set to nothing
-    }
-    //Backup hand
-    
-    //Update Coins for Buy
-    updateCoins(currentPlayer, state, 5);
-    x = 1;//Condition to loop on
-    while( x == 1) {//Buy one card
-        if (supplyCount(choice1, state) <= 0){
-            if (DEBUG)
-                printf("None of that card left, sorry!\n");
-            
-            if (DEBUG){
-                printf("Cards Left: %d\n", supplyCount(choice1, state));
-            }
-        }
-        else if (state->coins < getCost(choice1)){
-            printf("That card is too expensive!\n");
-            
-            if (DEBUG){
-                printf("Coins: %d < %d\n", state->coins, getCost(choice1));
-            }
-        }
-        else{
-            
-            if (DEBUG){
-                printf("Deck Count: %d\n", state->handCount[currentPlayer] + state->deckCount[currentPlayer] + state->discardCount[currentPlayer]);
-            }
-            //*BUG INTRODUCED*
-            gainCard(0, state, 0, currentPlayer);//Gain the card
-            x = 0;//No more buying cards
-            
-            if (DEBUG){
-                printf("Deck Count: %d\n", state->handCount[currentPlayer] + state->deckCount[currentPlayer] + state->discardCount[currentPlayer]);
-            }
-            
-        }
-    }     
-    
-    //Reset Hand
-    for (i = 0; i <= state->handCount[currentPlayer]; i++){
-        state->hand[currentPlayer][i] = temphand[i];
-        temphand[i] = -1;
-    }
-    //Reset Hand
-}
-
-void villageRefactor(int currentPlayer, struct gameState *state, int handPos)
-{
-    //+1 Card
-    drawCard(currentPlayer, state);
-    
-    //-1 Actions
-    //*BUG INTRODUCED*
-    state->numActions = state->numActions - 1;
-    
-    //discard played card from hand
-    discardCard(handPos, currentPlayer, state, 0);
-}
-
-void greatHallRefactor(int currentPlayer, struct gameState *state, int handPos)
-{
-    //+1 Card
-    drawCard(currentPlayer, state);
-    
-    //+1 Actions
-    state->numActions++;
-    
-    //discard card from hand
-    discardCard(handPos, currentPlayer, state, 0);
-}
-
-//End of REFACTOR section
-
 int compare(const void* a, const void* b) {
   if (*(int*)a > *(int*)b)
     return 1;
@@ -787,7 +656,7 @@ int cardEffect(int card, int choice1, int choice2, int choice3, struct gameState
   int tributeRevealedCards[2] = {-1, -1};
   int temphand[MAX_HAND];// moved above the if statement
   int drawntreasure=0;
-  //int cardDrawn;  //Silence Warning
+  int cardDrawn;
   int z = 0;// this is the counter for the temp hand
   if (nextPlayer > (state->numPlayers - 1)){
     nextPlayer = 0;
@@ -798,10 +667,24 @@ int cardEffect(int card, int choice1, int choice2, int choice3, struct gameState
   switch( card ) 
     {
     case adventurer:
-    
-    //Call to refactored function
-    adventurerRefactor(drawntreasure, state, currentPlayer, temphand, z);
-    
+      while(drawntreasure<2){
+	if (state->deckCount[currentPlayer] <1){//if the deck is empty we need to shuffle discard and add to deck
+	  shuffle(currentPlayer, state);
+	}
+	drawCard(currentPlayer, state);
+	cardDrawn = state->hand[currentPlayer][state->handCount[currentPlayer]-1];//top card of hand is most recently drawn card.
+	if (cardDrawn == copper || cardDrawn == silver || cardDrawn == gold)
+	  drawntreasure++;
+	else{
+	  temphand[z]=cardDrawn;
+	  state->handCount[currentPlayer]--; //this should just remove the top card (the most recently drawn one).
+	  z++;
+	}
+      }
+      while(z-1>=0){
+	state->discard[currentPlayer][state->discardCount[currentPlayer]++]=temphand[z-1]; // discard all cards in play that have been drawn
+	z=z-1;
+      }
       return 0;
 			
     case council_room:
@@ -829,12 +712,56 @@ int cardEffect(int card, int choice1, int choice2, int choice3, struct gameState
       return 0;
 			
     case feast:
-      i = 0; //Initialize for passing into refactored function
-      x = 1; //Initialize for passing into refactored function
-    
-     //Call to refactored function
-     feastRefactor(currentPlayer, state, i, temphand, x, choice1);
-            
+      //gain card with cost up to 5
+      //Backup hand
+      for (i = 0; i <= state->handCount[currentPlayer]; i++){
+	temphand[i] = state->hand[currentPlayer][i];//Backup card
+	state->hand[currentPlayer][i] = -1;//Set to nothing
+      }
+      //Backup hand
+
+      //Update Coins for Buy
+      updateCoins(currentPlayer, state, 5);
+      x = 1;//Condition to loop on
+      while( x == 1) {//Buy one card
+	if (supplyCount(choice1, state) <= 0){
+	  if (DEBUG)
+	    printf("None of that card left, sorry!\n");
+
+	  if (DEBUG){
+	    printf("Cards Left: %d\n", supplyCount(choice1, state));
+	  }
+	}
+	else if (state->coins < getCost(choice1)){
+	  printf("That card is too expensive!\n");
+
+	  if (DEBUG){
+	    printf("Coins: %d < %d\n", state->coins, getCost(choice1));
+	  }
+	}
+	else{
+
+	  if (DEBUG){
+	    printf("Deck Count: %d\n", state->handCount[currentPlayer] + state->deckCount[currentPlayer] + state->discardCount[currentPlayer]);
+	  }
+
+	  gainCard(choice1, state, 0, currentPlayer);//Gain the card
+	  x = 0;//No more buying cards
+
+	  if (DEBUG){
+	    printf("Deck Count: %d\n", state->handCount[currentPlayer] + state->deckCount[currentPlayer] + state->discardCount[currentPlayer]);
+	  }
+
+	}
+      }     
+
+      //Reset Hand
+      for (i = 0; i <= state->handCount[currentPlayer]; i++){
+	state->hand[currentPlayer][i] = temphand[i];
+	temphand[i] = -1;
+      }
+      //Reset Hand
+      			
       return 0;
 			
     case gardens:
@@ -902,19 +829,25 @@ int cardEffect(int card, int choice1, int choice2, int choice3, struct gameState
       return 0;
 		
     case smithy:
-       
-        i = 0; //Initialize variable for passing into refactoring function.
-            
-        //Call to refactored function
-       smithyRefactor(currentPlayer, state, i, handPos);
-            
+      //+3 Cards
+      for (i = 0; i < 3; i++)
+	{
+	  drawCard(currentPlayer, state);
+	}
+			
+      //discard card from hand
+      discardCard(handPos, currentPlayer, state, 0);
       return 0;
 		
     case village:
-     
-    //Call to refactored function
-    villageRefactor(currentPlayer, state, handPos);
-            
+      //+1 Card
+      drawCard(currentPlayer, state);
+			
+      //+2 Actions
+      state->numActions = state->numActions + 2;
+			
+      //discard played card from hand
+      discardCard(handPos, currentPlayer, state, 0);
       return 0;
 		
     case baron:
@@ -969,10 +902,14 @@ int cardEffect(int card, int choice1, int choice2, int choice3, struct gameState
       return 0;
 		
     case great_hall:
-     
-    //Call to refactored function
-    greatHallRefactor(currentPlayer, state, handPos);
-            
+      //+1 Card
+      drawCard(currentPlayer, state);
+			
+      //+1 Actions
+      state->numActions++;
+			
+      //discard card from hand
+      discardCard(handPos, currentPlayer, state, 0);
       return 0;
 		
     case minion:
@@ -1390,8 +1327,6 @@ int updateCoins(int player, struct gameState *state, int bonus)
 
   return 0;
 }
-
-
 
 
 //end of dominion.c
