+++ dominion/player.c
@@ -27,8 +27,6 @@ int main2(int argc, char *argv[]) {
 	initializeGame(2,k,1,&g);
 	printf ("SUCCESSFUL INIT\n");
 	getchar();
-    
-    return 0; //Silenced Warning
 }
 
 int main(int argc, char* argv[]) {
