+++ dominion/testBuyCard.c
@@ -4,8 +4,6 @@
 #include <stdio.h>
 #include <assert.h>
 #include "rngs.h"
-#include <math.h> //Added to silence warnings
-#include <stdlib.h> //Added to silence warnings
 
 #define DEBUG 0
 #define NOISY_TEST 1
