+++ projects/proulxp/dominion/testDrawCard.c
@@ -4,6 +4,8 @@
 #include <stdio.h>
 #include <assert.h>
 #include "rngs.h"
+#include <math.h> //Added to silence warnings
+#include <stdlib.h> //Added to silence warnings
 
 #define DEBUG 0
 #define NOISY_TEST 1
@@ -37,14 +39,20 @@ int checkDrawCard(int p, struct gameState *post) {
   assert (r == 0);
 
   assert(memcmp(&pre, post, sizeof(struct gameState)) == 0);
+    
+    return 0; //Silence warning
 }
 
 int main () {
-
-  int i, n, r, p, deckCount, discardCount, handCount;
-
+   
+    //Removed r, deckCount, discardCount, and handCount to silence warnings
+  int i, n, p;
+  
+    //Silence Warnings
+    /*
   int k[10] = {adventurer, council_room, feast, gardens, mine,
 	       remodel, smithy, village, baron, great_hall};
+     */
 
   struct gameState G;
 
@@ -69,7 +77,9 @@ int main () {
   printf ("ALL TESTS OK\n");
 
   exit(0);
-
+   
+    //Silence warning of unexecuted code
+  /*
   printf ("SIMPLE FIXED TESTS.\n");
   for (p = 0; p < 2; p++) {
     for (deckCount = 0; deckCount < 5; deckCount++) {
@@ -88,6 +98,6 @@ int main () {
       }
     }
   }
-
+  */
   return 0;
 }
